requirements.txt
python-telegram-bot>=20.0
anthropic>=0.40.0
openai>=1.0.0
APScheduler>=3.10.0,<4.0.0
python-dotenv>=1.0.0
python-docx>=1.1.0
httpx>=0.27.0
sentence-transformers>=3.0.0
